requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-heroku==0.3.1
djangorestframework==3.15.2
docx2pdf==0.1.8
fire==0.7.0
fonttools==4.56.0
gunicorn==23.0.0
idna==3.10
lxml==5.3.1
numpy==2.2.3
opencv-python-headless==4.11.0.86
packaging==24.2
pdf2docx==0.5.8
pillow==11.1.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMuPDF==1.25.3
python-docx==1.1.2
pywin32==308
qrcode==8.0
requests==2.32.3
sqlparse==0.5.3
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
